+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/transactions/layered/LayeredPendingTransactions.java
@@ -433,7 +433,7 @@ public class LayeredPendingTransactions implements PendingTransactions {
   }
 
   @Override
-  public synchronized void manageBlockAdded(
+  public void manageBlockAdded(
       final BlockHeader blockHeader,
       final List<Transaction> confirmedTransactions,
       final List<Transaction> reorgTransactions,
@@ -447,19 +447,21 @@ public class LayeredPendingTransactions implements PendingTransactions {
 
     final var reorgNonceRangeBySender = nonceRangeBySender(reorgTransactions);
 
-    try {
-      prioritizedTransactions.blockAdded(feeMarket, blockHeader, maxConfirmedNonceBySender);
-    } catch (final Throwable throwable) {
-      LOG.warn(
-          "Unexpected error {} when managing added block {}, maxNonceBySender {}, reorgNonceRangeBySender {}",
-          throwable,
-          blockHeader.toLogString(),
-          maxConfirmedNonceBySender,
-          reorgTransactions);
-      LOG.warn("Stack trace", throwable);
-    }
+    synchronized (this) {
+      try {
+        prioritizedTransactions.blockAdded(feeMarket, blockHeader, maxConfirmedNonceBySender);
+      } catch (final Throwable throwable) {
+        LOG.warn(
+            "Unexpected error {} when managing added block {}, maxNonceBySender {}, reorgNonceRangeBySender {}",
+            throwable,
+            blockHeader.toLogString(),
+            maxConfirmedNonceBySender,
+            reorgTransactions);
+        LOG.warn("Stack trace", throwable);
+      }
 
-    logBlockHeaderForReplay(blockHeader, maxConfirmedNonceBySender, reorgNonceRangeBySender);
+      logBlockHeaderForReplay(blockHeader, maxConfirmedNonceBySender, reorgNonceRangeBySender);
+    }
   }
 
   private void logBlockHeaderForReplay(
@@ -498,10 +500,25 @@ public class LayeredPendingTransactions implements PendingTransactions {
   }
 
   private Map<Address, Long> maxNonceBySender(final List<Transaction> confirmedTransactions) {
+    record SenderNonce(Address sender, long nonce) {}
+
     return confirmedTransactions.stream()
+        .<SenderNonce>mapMulti(
+            (transaction, consumer) -> {
+              // always consider the sender
+              consumer.accept(new SenderNonce(transaction.getSender(), transaction.getNonce()));
+
+              // and if a code delegation tx also the authorities
+              if (transaction.getType().supportsDelegateCode()) {
+                transaction.getCodeDelegationList().get().stream()
+                    .map(cd -> cd.authorizer().map(address -> new SenderNonce(address, cd.nonce())))
+                    .filter(Optional::isPresent)
+                    .map(Optional::get)
+                    .forEach(consumer);
+              }
+            })
         .collect(
-            groupingBy(
-                Transaction::getSender, mapping(Transaction::getNonce, reducing(0L, Math::max))));
+            groupingBy(SenderNonce::sender, mapping(SenderNonce::nonce, reducing(0L, Math::max))));
   }
 
   private Map<Address, LongRange> nonceRangeBySender(