+++ CHANGELOG.md
@@ -16,10 +16,9 @@
   - Proof of Work consensus
   - Fast Sync
 
-
-
 ### Additions and Improvements
 - Retrieve all transaction receipts for a block in one request [#6646](https://github.com/hyperledger/besu/pull/6646)
+- Add support for EIP-7702 transaction in the txpool [#8018](https://github.com/hyperledger/besu/pull/8018) [#7984](https://github.com/hyperledger/besu/pull/7984)
 
 ### Bug fixes